We need to take DevicePixelRatio into account when on high res screens
We need to make sure that the maxHeight is calculate by splitting it by
devicePixelRatio since size() in Qt is the *real* size and not the size
/ dpr, so when calculating the different scales, make sure to use dpr
to split max height and get a scaled value.

Also trigger the scale update when switching into fullscreen and back
to windowed mode.